revert:fix(common): detect data: and blob: inputs in NgOptimizedImage
The `src` attribute supports blob and data URLs, and allowing the same
behavior for `ngSrc` doesn't have any visible negative
effects from my initial tests. Since there is no documented rationale
for this, either in the original commit or official docs, I am assuming
this is not a needed restriction and can be considered a regression.

NOTE: If this restriction _was_ to be required, both documentation and
linters should be updated accordingly. Various IDE's now assume Angular
projects should use `ngSrc` as a replacement for `src` now, and will
inform users of that.

This reverts commit 801daf8.